+++ src/hooks/api/talk-pick/useSaveTempTalkPickMutation.ts
@@ -1,26 +1,14 @@
-/* eslint-disable @typescript-eslint/no-unused-vars */
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 import { HTTP_STATUS_CODE } from '@/constants/api';
 import { postTempTalkPick } from '@/api/talk-pick';
 import { NewTalkPick } from '@/types/talk-pick';
-import { useState } from 'react';
-import { useNavigate } from 'react-router-dom';
 import { AxiosErrorResponse } from '@/api/interceptor';
 import { ERROR } from '@/constants/message';
 
-export const useSaveTempTalkPickMutation = () => {
+export const useSaveTempTalkPickMutation = (
+  showToastModal: (message: string) => () => void,
+) => {
   const queryClient = useQueryClient();
-  const [saveSuccess, setSaveSuccess] = useState<boolean>(false);
-  const [isError, setIsError] = useState<boolean>(false);
-  const [errorMessage, setErrorMessage] = useState<string | undefined>(
-    undefined,
-  );
-  const navigate = useNavigate();
-
-  const setCreateError = (message: string) => {
-    setIsError(true);
-    setErrorMessage(message);
-  };
 
   const mutation = useMutation({
     mutationFn: (data: NewTalkPick) => postTempTalkPick(data),
@@ -29,17 +17,14 @@ export const useSaveTempTalkPickMutation = () => {
         queryKey: ['tempTalkPick'],
       });
 
-      setSaveSuccess(true);
+      showToastModal('임시저장 완료!');
     },
     onError: (err: AxiosErrorResponse) => {
       if (err.status === HTTP_STATUS_CODE.UNAUTHORIZED) {
-        setCreateError(ERROR.SAVE.FAIL);
+        showToastModal(ERROR.SAVE.FAIL);
       }
     },
   });
 
-  return {
-    ...mutation,
-    saveSuccess,
-  };
+  return { ...mutation };
 };
